Fix:Gemini config, 컨트롤러, 서비스 병합, 트렌젝션 타입 변경
