+++ src/main/java/org/solcation/solcation_be/config/GeminiConfig.java
@@ -1,6 +1,6 @@
 package org.solcation.solcation_be.config;
 
-import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.MediaType;
@@ -11,26 +11,36 @@ import org.springframework.web.util.UriComponentsBuilder;
 import java.net.URI;
 
 @Configuration
-@RequiredArgsConstructor
 public class GeminiConfig {
 
-    private final GeminiProperties geminiProperties;
+    @Value("${gemini.api-key:}")
+    private String apiKey;
+
+    @Value("${gemini.model:gemini-2.5-flash}")
+    private String model;
+
+    @Value("${gemini.endpoint:https://generativelanguage.googleapis.com/v1beta/models}")
+    private String endpoint;
 
     @Bean
     public WebClient geminiWebClient(WebClient.Builder builder) {
-        if (geminiProperties.getApiKey() == null || geminiProperties.getApiKey().isBlank()) {
+        if (apiKey == null || apiKey.isBlank()) {
             throw new IllegalStateException("GOOGLE_API_KEY 환경변수가 비어있어");
         }
         return builder
-                .baseUrl(geminiProperties.getEndpoint())
+                .baseUrl(endpoint)
                 .defaultHeaders(h -> h.setContentType(MediaType.APPLICATION_JSON))
                 .filter((request, next) -> {
                     URI withKey = UriComponentsBuilder.fromUri(request.url())
-                            .queryParam("key", geminiProperties.getApiKey())
+                            .queryParam("key", apiKey)
                             .build(true).toUri();
                     ClientRequest newReq = ClientRequest.from(request).url(withKey).build();
                     return next.exchange(newReq);
                 })
                 .build();
+    }
+
+    public String geminiModel() {
+        return model;
     }
 }
