--- src/main/java/org/solcation/solcation_be/config/GeminiProperties.java
@@ -1,16 +1,0 @@
-package org.solcation.solcation_be.config;
-
-import lombok.Getter;
-import lombok.Setter;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.stereotype.Component;
-
-@Getter
-@Setter
-@Component
-@ConfigurationProperties(prefix = "gemini")
-public class GeminiProperties {
-    private String apiKey;
-    private String model;
-    private String endpoint;
-}
